Fixes to Plugin
Filled in definition for doRegenerate() function, and changed doSubmit() to only take two arguments, target and which. The query data can be accessed from the target. Also fixed issue in html form where the Instructions div was not collapsing text. Plugin now replaces sql search text with desired values, submits the proper query, and returns the results in the correct format.